backend/requirements.txt
asgiref==3.7.2
Django==3.2.9
django-cors-headers==3.13.0
djangorestframework==3.12.4
psycopg2-binary==2.9.3
sqlparse==0.4.4
typing_extensions==4.7.1
